Simplified file structure
